Refactored and cleaned up the idmapper directory; it's been so changed by evennia that it makes little sense to keep the original structure (which was intended for adding an app into). Resolved the unittest errors with the idmapper.
